jwt add .....

+++ src/main/java/com/airportSystem/airport_system/Controller/FlightController.java
@@ -4,19 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.airportSystem.airport_system.Dao.SeatRepository;
-import com.airportSystem.airport_system.Entities.FlightDto;
 import com.airportSystem.airport_system.Entities.Flights;
 import com.airportSystem.airport_system.Entities.Seat;
 import com.airportSystem.airport_system.Entities.SeatKey;
@@ -32,18 +28,12 @@ public class FlightController {
     @Autowired
     private SeatRepository seatRepository;
 
-    @GetMapping("/flight/{id}")
-    public Flights getflightById(@PathVariable("id") Long id) {
-        return flightService.getFlightById(id).orElse(null);
-    }
-
-    @GetMapping("/allFlights")
-    public Page<FlightDto> getAllFlights(Pageable pageable) {
-        return flightService.getAllFlights(pageable);
-    }
-
     @PostMapping("/addFlight")
     public String addFlight(@RequestBody Flights flight) {
+        String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        if (!role.contains("ADMIN")) {
+            return "Unauthorized";
+        }
         Flights savedFlight = flightService.saveFlight(flight);
         char[] columns = { 'A', 'B', 'C', 'D', 'E', 'F' };
 
@@ -68,22 +58,13 @@ public class FlightController {
 
     @DeleteMapping("/deleteFlight/{id}")
     public String deleteFlight(@PathVariable("id") Long id) {
+        String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        if (!role.contains("ADMIN")) {
+            return "Unauthorized";
+        }
         return flightService.deleteFlight(id);
     }
 
-    @GetMapping("/{id}/seats")
-    public List<List<Boolean>> getSeatsByClass(@PathVariable("id") Long id) {
-        return flightService.getSeatsByClass(id);
-    }
-
-    @GetMapping("/flightUsingLocation")
-    public List<FlightDto> returnFlightByOriginAndDestination(@RequestParam("originCountry") String oCountry,
-            @RequestParam("originState") String oState,
-            @RequestParam("originCity") String ocity,
-            @RequestParam("destinationCountry") String dCountry,
-            @RequestParam("destinationState") String dState,
-            @RequestParam("destinationCity") String dCity) {
-        return flightService.FlightByOriginAndDestination(oCountry, oState, ocity, dCountry, dState, dCity);
-    }
+    
 
 }
